Added SX Bet adapter for SX Staking and TVL via escrow
